# api-docs/swagger/openapi.yaml
...
openapi: 3.0.3
info:
  title: School Management API
  version: 1.0.0
  description: |
    A comprehensive REST API for managing school operations including classes, students, teachers, subjects, lesson schedules, and attendance tracking.
    
    ## Features
    - **Class Management**: Create and manage school classes with grade levels and letters
    - **Student Management**: Handle student enrollment and information
    - **Teacher Management**: Manage teacher profiles and assignments
    - **Subject Management**: Organize school subjects and curriculum
    - **Lesson Scheduling**: Plan and manage weekly lesson schedules
    - **Lesson Logging**: Track actual lessons that occurred
    - **Attendance & Grading**: Record student attendance and grades
    - **Teacher Assignments**: Link teachers to specific subjects
    
    ## Data Model
    The API follows a relational data model where:
    - Students belong to Classes
    - Teachers are assigned to Subjects via TeacherAssignments
    - LessonSchedules define when subjects are taught to classes
    - LessonLogs record actual lessons that happened
    - StudentLessons track individual student participation, grades, and attendance
servers:
  - url: https://d5ds9jru50sk4al4aiaa.svoluuab.apigw.yandexcloud.net
    description: Local development server
...
paths:
  /api/v1/classes:
    get:
      tags:
        - Classes
      summary: List all classes
      operationId: listClasses
      parameters:
        - in: query
          name: grade
          schema:
            type: integer
            minimum: 1
            maximum: 12
          description: Filter by grade level
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
      responses:
        "200":
          description: A list of classes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Class"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Classes
      summary: Create a new class
      operationId: createClass
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  "/api/v1/classes/{id}":
    get:
      tags:
        - Classes
      summary: Get class by ID
      operationId: getClass
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Classes
      summary: Update class
      operationId: updateClass
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "200":
          description: Class updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Classes
      summary: Delete class
      operationId: deleteClass
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Class deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/lesson-logs:
    get:
      tags:
        - LessonLogs
      summary: List all lesson logs
      operationId: listLessonLogs
      parameters:
        - in: query
          name: class_id
          schema:
            type: integer
          description: Filter by class ID
        - in: query
          name: teacher_id
          schema:
            type: integer
          description: Filter by teacher ID
        - in: query
          name: subject_id
          schema:
            type: integer
          description: Filter by subject ID
        - in: query
          name: date_from
          schema:
            type: string
            format: date
          description: Filter from date (YYYY-MM-DD)
        - in: query
          name: date_to
          schema:
            type: string
            format: date
          description: Filter to date (YYYY-MM-DD)
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
      responses:
        "200":
          description: A list of lesson logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/LessonLog"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - LessonLogs
      summary: Create a new lesson log
      operationId: createLessonLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonLogInput"
      responses:
        "201":
          description: Lesson log created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonLog"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  "/api/v1/lesson-logs/{id}":
    get:
      tags:
        - LessonLogs
      summary: Get lesson log by ID
      operationId: getLessonLog
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson log found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonLog"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - LessonLogs
      summary: Update lesson log
      operationId: updateLessonLog
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonLogInput"
      responses:
        "200":
          description: Lesson log updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonLog"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - LessonLogs
      summary: Delete lesson log
      operationId: deleteLessonLog
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Lesson log deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/attendance-statuses:
    get:
      tags:
        - AttendanceStatuses
      summary: List all attendance statuses
      operationId: listAttendanceStatuses
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
      responses:
        "200":
          description: A list of attendance statuses
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AttendanceStatus"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - AttendanceStatuses
      summary: Create a new attendance status
      operationId: createAttendanceStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceStatusInput"
      responses:
        "201":
          description: Attendance status created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  "/api/v1/attendance-statuses/{code}":
    get:
      tags:
        - AttendanceStatuses
      summary: Get attendance status by code
      operationId: getAttendanceStatus
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
            maxLength: 1
      responses:
        "200":
          description: Attendance status found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - AttendanceStatuses
      summary: Update attendance status
      operationId: updateAttendanceStatus
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
            maxLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceStatusInput"
      responses:
        "200":
          description: Attendance status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - AttendanceStatuses
      summary: Delete attendance status
      operationId: deleteAttendanceStatus
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
            maxLength: 1
      responses:
        "204":
          description: Attendance status deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  schemas:
...
    Class:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the class
        grade:
          type: integer
          minimum: 1
          maximum: 12
          description: Grade level (1-12)
        letter:
          type: string
          minLength: 1
          maxLength: 1
          pattern: "^[A-Z]$"
          description: Class letter (A-Z)
      required:
        - grade
        - letter
      example:
        id: 1
        grade: 10
        letter: "A"
...
    LessonLog:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the lesson log
        subject_id:
          type: integer
          description: ID of the subject
        date:
          type: string
          format: date
          description: Date when the lesson occurred (YYYY-MM-DD)
        number:
          type: integer
          minimum: 1
          maximum: 8
          description: Lesson number in the day (1-8)
        class_id:
          type: integer
          description: ID of the class
        teacher_id:
          type: integer
          description: ID of the teacher
      required:
        - subject_id
        - date
        - number
        - class_id
        - teacher_id
      example:
        id: 1
        subject_id: 1
        date: "2024-01-15"
        number: 1
        class_id: 1
        teacher_id: 1
...
    AttendanceStatus:
      type: object
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 1
          pattern: "^[PALES]$"
          description: Single character code for attendance status
        description:
          type: string
          minLength: 1
          maxLength: 50
          description: Human-readable description of the status
      required:
        - code
        - description
      example:
        code: "P"
        description: "Present"

    # Input Schemas (without IDs)
    ClassInput:
      type: object
      properties:
        grade:
          type: integer
          minimum: 1
          maximum: 12
          description: Grade level (1-12)
        letter:
          type: string
          minLength: 1
          maxLength: 1
          pattern: "^[A-Z]$"
          description: Class letter (A-Z)
      required:
        - grade
        - letter
...
    LessonLogInput:
      type: object
      properties:
        subject_id:
          type: integer
          description: ID of the subject
        date:
          type: string
          format: date
          description: Date when the lesson occurred (YYYY-MM-DD)
        number:
          type: integer
          minimum: 1
          maximum: 8
          description: Lesson number in the day (1-8)
        class_id:
          type: integer
          description: ID of the class
        teacher_id:
          type: integer
          description: ID of the teacher
      required:
        - subject_id
        - date
        - number
        - class_id
        - teacher_id
...
    AttendanceStatusInput:
      type: object
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 1
          pattern: "^[PALES]$"
          description: Single character code for attendance status
        description:
          type: string
          minLength: 1
          maxLength: 50
          description: Human-readable description of the status
      required:
        - code
        - description

    # Utility Schemas
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: Number of items per page
        total:
          type: integer
          minimum: 0
          description: Total number of items
        pages:
          type: integer
          minimum: 0
          description: Total number of pages
      required:
        - page
        - limit
        - total
        - pages
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
      required:
        - error
        - message

  # Reusable Response Definitions
  responses:
    BadRequest:
      description: Bad Request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "BadRequest"
            message: "Invalid input parameters"
            details:
              field: "grade"
              issue: "Must be between 1 and 12"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "NotFound"
            message: "Resource not found"
    UnprocessableEntity:
      description: Unprocessable Entity - Business logic validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "UnprocessableEntity"
            message: "Business logic validation failed"
            details:
              field: "teacher_id"
              issue: "Teacher is already assigned to this subject"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "InternalServerError"
            message: "An unexpected error occurred"